Multiple changes:
 - Tests split into "slow" and "fast".
 - Binary files are uploaded only once, at the beginning of slow tests, and
   deleted regardless of whether test time-out has been reached.
 - SAVE project named "Diktat-Integration" is now used by default.
 - The project name is externally configurable via the `save-cloud.project.name`
   Gradle or System property.
