+++ .github/workflows/test.yml
@@ -42,6 +42,7 @@ jobs:
           -Dsave-cloud.user=github@${{ secrets.SAVE_CLOUD_USER }}
           -Dsave-cloud.password=${{ secrets.SAVE_CLOUD_PASSWORD }}
           -Dsave-cloud.user.auth.source=github
+          -Dsave-cloud.project.name=Diktat-Integration
           -Dsave-cloud.contest.name=new1
           -Dsave-cloud.test.version=save-integration-test
           build
